fix(mobile): rename, move bb-245
