+++ apps/mobile/src/libs/components/infinite-pager/infinite-pager.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { default as RNInfinitePager } from "react-native-infinite-pager";
 
-import { PREVIOUS_INDEX_OFFSET } from "~/libs/constants/constants";
+import { NumericalValue } from "~/libs/enums/enums";
 import { type InfinitePagerImperativeApi } from "~/libs/types/types";
 
 import { getPageInterpolatorSlide } from "./libs/animations/animations";
@@ -9,19 +9,19 @@ import { ANIMATION_CONFIG } from "./libs/constants/constants";
 
 type Properties = {
 	infinitePagerReference: React.MutableRefObject<InfinitePagerImperativeApi | null>;
-	renderPageComponent: () => JSX.Element;
+	onPageRender: () => JSX.Element;
 };
 
 const InfinitePager = ({
 	infinitePagerReference,
-	renderPageComponent,
+	onPageRender,
 }: Properties): JSX.Element => {
 	return (
 		<RNInfinitePager
 			animationConfig={ANIMATION_CONFIG}
 			gesturesDisabled
-			pageBuffer={PREVIOUS_INDEX_OFFSET}
-			PageComponent={renderPageComponent}
+			pageBuffer={NumericalValue.ONE}
+			PageComponent={onPageRender}
 			pageInterpolator={getPageInterpolatorSlide}
 			ref={infinitePagerReference}
 		/>
