+++ apps/mobile/src/libs/types/types.ts
@@ -1,6 +1,6 @@
 export { type AsyncThunkConfig } from "./app/app";
 export { type IconName } from "./icon/icon";
-export { type InfinitePagerImperativeApi } from "./infinite-pager/types";
+export { type InfinitePagerImperativeApi } from "./infinite-pager/infinite-pager";
 export {
 	type BottomTabNavigationParameterList,
 	type NavigationScreenProperties,
